Add "gatekeeper mode" for ez DM-ing
Add a way to have a gatekeeper-only view that prompts with text to read
aloud. With this change, you can load
"http://localhost:8080/#gatekeeper" and see a view that will just have
text for the gatekeeper to read. When the gatekeeper triggers an attack,
a new piece of Redux middleware computes the text (using templating for
the player's name and the boss's name) and stores it in HTML5 Local
Storage. The Gatekeeper screen then listens to `storage` events and will
update its view when this occurs. This allows easy IPC without any
servers or sockets or similar -- the gatekeeper screen and the game
screen just need to be on the same origin.